AGS: Added AGSSnowRain plugin
